requirements.txt
openai
rich
mcp
python-dotenv
matplotlib
pandas
numpy
requests
textblob
docker
psutil
scikit-learn
cryptography
watchdog
PyYAML
PyPDF2
beautifulsoup4
pydantic
boto3
redis
PyJWT
Jinja2
dnspython
